[frontend/backend] Be able to associate files at creation in all entities (#2561)
