+++ opencti-platform/opencti-front/src/private/components/analysis/notes/NoteCreation.tsx
@@ -35,6 +35,8 @@ import { NotesLinesPaginationQuery$variables } from './__generated__/NotesLinesP
 import SliderField from '../../../../components/SliderField';
 import { ExternalReferencesField } from '../../common/form/ExternalReferencesField';
 import { useSchemaCreationValidation } from '../../../../utils/hooks/useEntitySettings';
+import { NoteCreationMutation$variables } from './__generated__/NoteCreationMutation.graphql';
+import { SimpleFileUpload } from 'formik-mui';
 
 const useStyles = makeStyles<Theme>((theme) => ({
   drawerPaper: {
@@ -122,6 +124,7 @@ interface NoteAddInput {
   objectMarking: Option[];
   objectLabel: Option[];
   externalReferences: { value: string }[];
+  file: File | undefined
 }
 
 interface NoteCreationProps {
@@ -180,6 +183,7 @@ export const NoteCreationForm: FunctionComponent<NoteFormProps> = ({
     objectMarking: defaultMarkingDefinitions ?? [],
     objectLabel: [],
     externalReferences: [],
+    file: undefined,
   };
   const [commit] = userIsKnowledgeEditor
     ? useMutation(noteCreationMutation)
@@ -188,7 +192,7 @@ export const NoteCreationForm: FunctionComponent<NoteFormProps> = ({
     values,
     { setSubmitting, resetForm },
   ) => {
-    const finalValues = {
+    const finalValues: NoteCreationMutation$variables['input'] = {
       created: values.created,
       attribute_abstract: values.attribute_abstract,
       content: values.content,
@@ -200,6 +204,9 @@ export const NoteCreationForm: FunctionComponent<NoteFormProps> = ({
       objectLabel: values.objectLabel.map((v) => v.value),
       externalReferences: values.externalReferences.map(({ value }) => value),
     };
+    if (values.file) {
+      finalValues.file = values.file;
+    }
     if (!userIsKnowledgeEditor) {
       delete finalValues.createdBy;
     }
@@ -299,6 +306,15 @@ export const NoteCreationForm: FunctionComponent<NoteFormProps> = ({
             style={fieldSpacingContainerStyle}
             setFieldValue={setFieldValue}
             values={values.externalReferences}
+          />
+          <Field
+            component={SimpleFileUpload}
+            name="file"
+            label={t('Associated file')}
+            FormControlProps={{ style: { marginTop: 20, width: '100%' } }}
+            InputLabelProps={{ fullWidth: true, variant: 'standard' }}
+            InputProps={{ fullWidth: true, variant: 'standard' }}
+            fullWidth={true}
           />
           <div className={classes.buttons}>
             <Button
