+++ opencti-platform/opencti-front/src/private/components/analysis/opinions/OpinionCreation.tsx
@@ -11,6 +11,7 @@ import { Add, Close } from '@mui/icons-material';
 import makeStyles from '@mui/styles/makeStyles';
 import { FormikConfig, FormikHelpers } from 'formik/dist/types';
 import { RecordSourceSelectorProxy } from 'relay-runtime';
+import { SimpleFileUpload } from 'formik-mui';
 import { handleErrorInForm } from '../../../../relay/environment';
 import { useFormatter } from '../../../../components/i18n';
 import ObjectMarkingField from '../../common/form/ObjectMarkingField';
@@ -29,6 +30,8 @@ import { useSchemaCreationValidation } from '../../../../utils/hooks/useEntitySe
 import useGranted, {
   KNOWLEDGE_KNUPDATE,
 } from '../../../../utils/hooks/useGranted';
+import { OpinionCreationMutation$variables } from './__generated__/OpinionCreationMutation.graphql';
+import { OpinionCreationUserMutation$variables } from './__generated__/OpinionCreationUserMutation.graphql';
 
 const useStyles = makeStyles<Theme>((theme) => ({
   drawerPaper: {
@@ -113,7 +116,9 @@ interface OpinionAddInput {
   objectMarking: Option[];
   objectLabel: Option[];
   externalReferences: { value: string }[];
+  file: File | undefined
 }
+
 
 interface OpinionCreationProps {
   paginationOptions: OpinionsLinesPaginationQuery$variables;
@@ -156,6 +161,7 @@ export const OpinionCreationFormKnowledgeEditor: FunctionComponent<OpinionFormPr
     objectMarking: defaultMarkingDefinitions ?? [],
     objectLabel: [],
     externalReferences: [],
+    file: undefined,
   };
 
   const [commit] = useMutation(opinionCreationMutation);
@@ -163,7 +169,7 @@ export const OpinionCreationFormKnowledgeEditor: FunctionComponent<OpinionFormPr
     values: OpinionAddInput,
     { setSubmitting, setErrors, resetForm }: FormikHelpers<OpinionAddInput>,
   ) => {
-    const finalValues = {
+    const finalValues: OpinionCreationMutation$variables['input'] = {
       opinion: values.opinion,
       explanation: values.explanation,
       confidence: parseInt(String(values.confidence), 10),
@@ -172,6 +178,9 @@ export const OpinionCreationFormKnowledgeEditor: FunctionComponent<OpinionFormPr
       objectLabel: values.objectLabel.map((v) => v.value),
       externalReferences: values.externalReferences.map(({ value }) => value),
     };
+    if (values.file) {
+      finalValues.file = values.file;
+    }
     commit({
       variables: {
         input: finalValues,
@@ -246,6 +255,15 @@ export const OpinionCreationFormKnowledgeEditor: FunctionComponent<OpinionFormPr
             setFieldValue={setFieldValue}
             values={values.externalReferences}
           />
+          <Field
+            component={SimpleFileUpload}
+            name="file"
+            label={t('Associated file')}
+            FormControlProps={{ style: { marginTop: 20, width: '100%' } }}
+            InputLabelProps={{ fullWidth: true, variant: 'standard' }}
+            InputProps={{ fullWidth: true, variant: 'standard' }}
+            fullWidth={true}
+          />
           <div className={classes.buttons}>
             <Button
               variant="contained"
@@ -299,6 +317,7 @@ export const OpinionCreationFormKnowledgeParticipant: FunctionComponent<OpinionF
     objectMarking: defaultMarkingDefinitions ?? [],
     objectLabel: [],
     externalReferences: [],
+    file: undefined,
   };
 
   const [commit] = useMutation(opinionCreationUserMutation);
@@ -306,7 +325,7 @@ export const OpinionCreationFormKnowledgeParticipant: FunctionComponent<OpinionF
     values: OpinionAddInput,
     { setSubmitting, setErrors, resetForm }: FormikHelpers<OpinionAddInput>,
   ) => {
-    const finalValues = {
+    const finalValues: OpinionCreationMutation$variables['input'] = {
       opinion: values.opinion,
       explanation: values.explanation,
       confidence: parseInt(String(values.confidence), 10),
@@ -315,6 +334,9 @@ export const OpinionCreationFormKnowledgeParticipant: FunctionComponent<OpinionF
       objectLabel: values.objectLabel.map((v) => v.value),
       externalReferences: values.externalReferences.map(({ value }) => value),
     };
+    if (values.file) {
+      finalValues.file = values.file;
+    }
     commit({
       variables: {
         input: finalValues,
@@ -383,6 +405,15 @@ export const OpinionCreationFormKnowledgeParticipant: FunctionComponent<OpinionF
             style={fieldSpacingContainerStyle}
             setFieldValue={setFieldValue}
             values={values.externalReferences}
+          />
+          <Field
+            component={SimpleFileUpload}
+            name="file"
+            label={t('Associated file')}
+            FormControlProps={{ style: { marginTop: 20, width: '100%' } }}
+            InputLabelProps={{ fullWidth: true, variant: 'standard' }}
+            InputProps={{ fullWidth: true, variant: 'standard' }}
+            fullWidth={true}
           />
           <div className={classes.buttons}>
             <Button
