+++ opencti-platform/opencti-front/src/private/components/arsenal/malwares/MalwareCreation.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React, { FunctionComponent, useState } from 'react';
 import { Field, Form, Formik } from 'formik';
 import Drawer from '@mui/material/Drawer';
 import Typography from '@mui/material/Typography';
@@ -7,11 +7,13 @@ import IconButton from '@mui/material/IconButton';
 import Fab from '@mui/material/Fab';
 import { Add, Close } from '@mui/icons-material';
 import * as Yup from 'yup';
-import { graphql } from 'react-relay';
-import * as R from 'ramda';
+import { graphql, useMutation } from 'react-relay';
 import makeStyles from '@mui/styles/makeStyles';
+import { SimpleFileUpload } from 'formik-mui';
+import { RecordSourceSelectorProxy } from 'relay-runtime';
+import { FormikConfig } from 'formik/dist/types';
 import { useFormatter } from '../../../../components/i18n';
-import { commitMutation, handleErrorInForm } from '../../../../relay/environment';
+import { handleErrorInForm } from '../../../../relay/environment';
 import TextField from '../../../../components/TextField';
 import CreatedByField from '../../common/form/CreatedByField';
 import ObjectLabelField from '../../common/form/ObjectLabelField';
@@ -25,8 +27,15 @@ import { insertNode } from '../../../../utils/store';
 import { fieldSpacingContainerStyle } from '../../../../utils/field';
 import { useSchemaCreationValidation } from '../../../../utils/hooks/useEntitySettings';
 import SwitchField from '../../../../components/SwitchField';
+import { Option } from '../../common/form/ReferenceField';
+import { Theme } from '../../../../components/Theme';
+import {
+  MalwareCreationMutation,
+  MalwareCreationMutation$variables,
+} from './__generated__/MalwareCreationMutation.graphql';
+import { MalwaresLinesPaginationQuery$variables } from './__generated__/MalwaresLinesPaginationQuery.graphql';
 
-const useStyles = makeStyles((theme) => ({
+const useStyles = makeStyles<Theme>((theme) => ({
   drawerPaper: {
     minHeight: '100vh',
     width: '50%',
@@ -79,13 +88,45 @@ const malwareMutation = graphql`
       parent_types
       architecture_execution_envs
       implementation_languages
+      malware_types
       ...MalwareCard_node
     }
   }
 `;
 
-export const MalwareCreationForm = ({ updater, onReset, onCompleted,
-  defaultConfidence, defaultCreatedBy, defaultMarkingDefinitions }) => {
+interface MalwareAddInput {
+  name: string
+  malware_types: Option[]
+  confidence: number
+  description: string
+  createdBy: Option | undefined
+  is_family: boolean
+  objectMarking: Option[]
+  killChainPhases: Option[],
+  objectLabel: Option[]
+  externalReferences: { value: string }[]
+  architecture_execution_envs: string[],
+  implementation_languages: string[],
+  file: File | undefined
+}
+
+interface MalwareFormProps {
+  updater: (store: RecordSourceSelectorProxy, key: string) => void
+  onReset?: () => void
+  onCompleted?: () => void;
+  defaultCreatedBy?: { value: string, label: string }
+  defaultMarkingDefinitions?: { value: string, label: string }[]
+  defaultConfidence?: number;
+}
+
+export const MalwareCreationForm: FunctionComponent<MalwareFormProps> = ({
+  updater,
+  onReset,
+  onCompleted,
+  defaultConfidence,
+  defaultCreatedBy,
+  defaultMarkingDefinitions,
+}) => {
   const classes = useStyles();
   const { t } = useFormatter();
   const basicShape = {
@@ -98,23 +139,46 @@ export const MalwareCreationForm = ({ updater, onReset, onCompleted,
     implementation_languages: Yup.array().nullable(),
   };
   const malwareValidator = useSchemaCreationValidation('Malware', basicShape);
-  const onSubmit = (values, { setSubmitting, setErrors, resetForm }) => {
-    const adaptedValues = R.evolve(
-      {
-        confidence: () => parseInt(values.confidence, 10),
-        createdBy: R.path(['value']),
-        killChainPhases: R.pluck('value'),
-        objectMarking: R.pluck('value'),
-        objectLabel: R.pluck('value'),
-        externalReferences: R.pluck('value'),
-        malware_types: R.pluck('value'),
-      },
-      values,
-    );
-    commitMutation({
-      mutation: malwareMutation,
+
+  const initialValues: MalwareAddInput = {
+    name: '',
+    malware_types: [],
+    confidence: defaultConfidence ?? 75,
+    description: '',
+    createdBy: defaultCreatedBy ?? '' as unknown as Option,
+    is_family: false,
+    objectMarking: defaultMarkingDefinitions ?? [],
+    killChainPhases: [],
+    objectLabel: [],
+    externalReferences: [],
+    architecture_execution_envs: [],
+    implementation_languages: [],
+    file: undefined,
+  };
+
+  const [commit] = useMutation<MalwareCreationMutation>(malwareMutation);
+
+  const onSubmit: FormikConfig<MalwareAddInput>['onSubmit'] = (values, { setSubmitting, setErrors, resetForm }) => {
+    const finalValues: MalwareCreationMutation$variables['input'] = {
+      name: values.name,
+      description: values.description,
+      is_family: values.is_family,
+      confidence: parseInt(String(values.confidence), 10),
+      malware_types: (values.malware_types ?? []).map(({ value }) => value),
+      killChainPhases: (values.killChainPhases ?? []).map(({ value }) => value),
+      architecture_execution_envs: values.architecture_execution_envs,
+      createdBy: values.createdBy?.value,
+      implementation_languages: values.implementation_languages,
+      objectMarking: values.objectMarking.map((v) => v.value),
+      objectLabel: values.objectLabel.map((v) => v.value),
+      externalReferences: values.externalReferences.map(({ value }) => value),
+    };
+    if (values.file) {
+      finalValues.file = values.file;
+    }
+    commit({
       variables: {
-        input: adaptedValues,
+        input: finalValues,
       },
       updater: (store) => {
         if (updater) {
@@ -125,7 +189,6 @@ export const MalwareCreationForm = ({ updater, onReset, onCompleted,
         handleErrorInForm(error, setErrors);
         setSubmitting(false);
       },
-      setSubmitting,
       onCompleted: () => {
         setSubmitting(false);
         resetForm();
@@ -137,20 +200,7 @@ export const MalwareCreationForm = ({ updater, onReset, onCompleted,
   };
 
   return <Formik
-      initialValues={{
-        name: '',
-        malware_types: [],
-        confidence: defaultConfidence ?? 75,
-        description: '',
-        createdBy: defaultCreatedBy ?? '',
-        is_family: false,
-        objectMarking: defaultMarkingDefinitions ?? [],
-        killChainPhases: [],
-        objectLabel: [],
-        externalReferences: [],
-        architecture_execution_envs: [],
-        implementation_languages: [],
-      }}
+      initialValues={initialValues}
       validationSchema={malwareValidator}
       onSubmit={onSubmit}
       onReset={() => {
@@ -248,6 +298,15 @@ export const MalwareCreationForm = ({ updater, onReset, onCompleted,
               setFieldValue={setFieldValue}
               values={values.externalReferences}
           />
+          <Field
+            component={SimpleFileUpload}
+            name="file"
+            label={t('Associated file')}
+            FormControlProps={{ style: { marginTop: 20, width: '100%' } }}
+            InputLabelProps={{ fullWidth: true, variant: 'standard' }}
+            InputProps={{ fullWidth: true, variant: 'standard' }}
+            fullWidth={true}
+          />
           <div className={classes.buttons}>
             <Button
                 variant="contained"
@@ -272,7 +331,7 @@ export const MalwareCreationForm = ({ updater, onReset, onCompleted,
   </Formik>;
 };
 
-const MalwareCreation = ({ paginationOptions }) => {
+const MalwareCreation = ({ paginationOptions }: { paginationOptions: MalwaresLinesPaginationQuery$variables }) => {
   const classes = useStyles();
   const { t } = useFormatter();
   const [open, setOpen] = useState(false);
@@ -280,7 +339,7 @@ const MalwareCreation = ({ paginationOptions }) => {
   const handleOpen = () => setOpen(true);
   const handleClose = () => setOpen(false);
   const onReset = () => handleClose();
-  const updater = (store) => insertNode(
+  const updater = (store: RecordSourceSelectorProxy) => insertNode(
     store,
     'Pagination_malwares',
     paginationOptions,
