+++ opencti-platform/opencti-front/src/private/components/arsenal/tools/ToolCreation.tsx
@@ -1,5 +1,4 @@
-import React, { useState } from 'react';
-import * as R from 'ramda';
+import React, { FunctionComponent, useState } from 'react';
 import { Field, Form, Formik } from 'formik';
 import Drawer from '@mui/material/Drawer';
 import Typography from '@mui/material/Typography';
@@ -8,10 +7,13 @@ import IconButton from '@mui/material/IconButton';
 import Fab from '@mui/material/Fab';
 import { Add, Close } from '@mui/icons-material';
 import * as Yup from 'yup';
-import { graphql } from 'react-relay';
+import { graphql, useMutation } from 'react-relay';
 import makeStyles from '@mui/styles/makeStyles';
+import { SimpleFileUpload } from 'formik-mui';
+import { RecordSourceSelectorProxy } from 'relay-runtime';
+import { FormikConfig } from 'formik/dist/types';
 import { useFormatter } from '../../../../components/i18n';
-import { commitMutation, handleErrorInForm } from '../../../../relay/environment';
+import { handleErrorInForm } from '../../../../relay/environment';
 import TextField from '../../../../components/TextField';
 import KillChainPhasesField from '../../common/form/KillChainPhasesField';
 import CreatedByField from '../../common/form/CreatedByField';
@@ -24,8 +26,12 @@ import { fieldSpacingContainerStyle } from '../../../../utils/field';
 import ConfidenceField from '../../common/form/ConfidenceField';
 import { useSchemaCreationValidation } from '../../../../utils/hooks/useEntitySettings';
 import { insertNode } from '../../../../utils/store';
+import { Option } from '../../common/form/ReferenceField';
+import { ToolsLinesPaginationQuery$variables } from './__generated__/ToolsLinesPaginationQuery.graphql';
+import { Theme } from '../../../../components/Theme';
+import { ToolCreationMutation, ToolCreationMutation$variables } from './__generated__/ToolCreationMutation.graphql';
 
-const useStyles = makeStyles((theme) => ({
+const useStyles = makeStyles<Theme>((theme) => ({
   drawerPaper: {
     minHeight: '100vh',
     width: '50%',
@@ -81,8 +87,36 @@ const toolMutation = graphql`
   }
 `;
 
-export const ToolCreationForm = ({ updater, onReset, onCompleted,
-  defaultConfidence, defaultCreatedBy, defaultMarkingDefinitions }) => {
+interface ToolAddInput {
+  name: string
+  description: string
+  createdBy: Option | undefined
+  objectMarking: Option[]
+  killChainPhases: Option[]
+  objectLabel: Option[]
+  externalReferences: { value: string }[]
+  tool_types: string[]
+  confidence: number
+  file: File | undefined
+}
+
+interface ToolFormProps {
+  updater: (store: RecordSourceSelectorProxy, key: string) => void
+  onReset?: () => void;
+  onCompleted?: () => void;
+  defaultCreatedBy?: { value: string, label: string }
+  defaultMarkingDefinitions?: { value: string, label: string }[]
+  defaultConfidence?: number;
+}
+
+export const ToolCreationForm: FunctionComponent<ToolFormProps> = ({
+  updater,
+  onReset,
+  onCompleted,
+  defaultConfidence,
+  defaultCreatedBy,
+  defaultMarkingDefinitions,
+}) => {
   const classes = useStyles();
   const { t } = useFormatter();
   const basicShape = {
@@ -92,18 +126,38 @@ export const ToolCreationForm = ({ updater, onReset, onCompleted,
     tool_types: Yup.array().nullable(),
   };
   const toolValidator = useSchemaCreationValidation('Tool', basicShape);
-  const onSubmit = (values, { setSubmitting, setErrors, resetForm }) => {
-    const finalValues = R.pipe(
-      R.assoc('confidence', parseInt(values.confidence, 10)),
-      R.assoc('createdBy', values.createdBy?.value),
-      R.assoc('objectMarking', R.pluck('value', values.objectMarking)),
-      R.assoc('killChainPhases', R.pluck('value', values.killChainPhases)),
-      R.assoc('objectLabel', R.pluck('value', values.objectLabel)),
-      R.assoc('externalReferences', R.pluck('value', values.externalReferences)),
-      R.assoc('tool_types', values.tool_types),
-    )(values);
-    commitMutation({
-      mutation: toolMutation,
+
+  const initialValues: ToolAddInput = {
+    name: '',
+    description: '',
+    createdBy: defaultCreatedBy ?? '' as unknown as Option,
+    objectMarking: defaultMarkingDefinitions ?? [],
+    killChainPhases: [],
+    objectLabel: [],
+    externalReferences: [],
+    tool_types: [],
+    confidence: defaultConfidence ?? 75,
+    file: undefined,
+  };
+
+  const [commit] = useMutation<ToolCreationMutation>(toolMutation);
+
+  const onSubmit: FormikConfig<ToolAddInput>['onSubmit'] = (values, { setSubmitting, setErrors, resetForm }) => {
+    const finalValues: ToolCreationMutation$variables['input'] = {
+      name: values.name,
+      description: values.description,
+      createdBy: values.createdBy?.value,
+      objectMarking: values.objectMarking.map((v) => v.value),
+      killChainPhases: (values.killChainPhases ?? []).map(({ value }) => value),
+      objectLabel: values.objectLabel.map((v) => v.value),
+      tool_types: values.tool_types,
+      confidence: parseInt(String(values.confidence), 10),
+      externalReferences: values.externalReferences.map(({ value }) => value),
+    };
+    if (values.file) {
+      finalValues.file = values.file;
+    }
+    commit({
       variables: {
         input: finalValues,
       },
@@ -116,7 +170,6 @@ export const ToolCreationForm = ({ updater, onReset, onCompleted,
         handleErrorInForm(error, setErrors);
         setSubmitting(false);
       },
-      setSubmitting,
       onCompleted: () => {
         setSubmitting(false);
         resetForm();
@@ -128,17 +181,7 @@ export const ToolCreationForm = ({ updater, onReset, onCompleted,
   };
 
   return <Formik
-      initialValues={{
-        name: '',
-        description: '',
-        createdBy: defaultCreatedBy ?? '',
-        objectMarking: defaultMarkingDefinitions ?? [],
-        killChainPhases: [],
-        objectLabel: [],
-        externalReferences: [],
-        tool_types: [],
-        confidence: defaultConfidence ?? 75,
-      }}
+      initialValues={initialValues}
       validationSchema={toolValidator}
       onSubmit={onSubmit}
       onReset={onReset}
@@ -205,6 +248,15 @@ export const ToolCreationForm = ({ updater, onReset, onCompleted,
               setFieldValue={setFieldValue}
               values={values.externalReferences}
           />
+          <Field
+            component={SimpleFileUpload}
+            name="file"
+            label={t('Associated file')}
+            FormControlProps={{ style: { marginTop: 20, width: '100%' } }}
+            InputLabelProps={{ fullWidth: true, variant: 'standard' }}
+            InputProps={{ fullWidth: true, variant: 'standard' }}
+            fullWidth={true}
+          />
           <div className={classes.buttons}>
             <Button
                 variant="contained"
@@ -229,14 +281,14 @@ export const ToolCreationForm = ({ updater, onReset, onCompleted,
   </Formik>;
 };
 
-const ToolCreation = ({ paginationOptions }) => {
+const ToolCreation = ({ paginationOptions }: { paginationOptions: ToolsLinesPaginationQuery$variables }) => {
   const classes = useStyles();
   const { t } = useFormatter();
   const [open, setOpen] = useState(false);
 
   const handleOpen = () => setOpen(true);
   const handleClose = () => setOpen(false);
-  const updater = (store) => insertNode(
+  const updater = (store: RecordSourceSelectorProxy) => insertNode(
     store,
     'Pagination_tools',
     paginationOptions,
