+++ opencti-platform/opencti-front/src/private/components/arsenal/vulnerabilities/VulnerabilityCreation.tsx
@@ -30,6 +30,8 @@ import ConfidenceField from '../../common/form/ConfidenceField';
 import SelectField from '../../../../components/SelectField';
 import { ExternalReferencesField } from '../../common/form/ExternalReferencesField';
 import { useSchemaCreationValidation } from '../../../../utils/hooks/useEntitySettings';
+import { SimpleFileUpload } from 'formik-mui';
+import { VulnerabilityCreationMutation$variables } from './__generated__/VulnerabilityCreationMutation.graphql';
 
 const useStyles = makeStyles<Theme>((theme) => ({
   drawerPaper: {
@@ -98,6 +100,7 @@ interface VulnerabilityAddInput {
   x_opencti_attack_vector: string
   confidence: number
   externalReferences: { value: string }[]
+  file: File | undefined
 }
 
 interface VulnerabilityCreationProps {
@@ -139,6 +142,7 @@ export const VulnerabilityCreationForm: FunctionComponent<VulnerabilityFormProps
     x_opencti_attack_vector: '',
     confidence: defaultConfidence ?? 75,
     externalReferences: [],
+    file: undefined,
   };
 
   const onSubmit: FormikConfig<VulnerabilityAddInput>['onSubmit'] = (
@@ -149,7 +153,7 @@ export const VulnerabilityCreationForm: FunctionComponent<VulnerabilityFormProps
       resetForm,
     }: FormikHelpers<VulnerabilityAddInput>,
   ) => {
-    const finalValues = {
+    const finalValues: VulnerabilityCreationMutation$variables['input'] = {
       name: values.name,
       description: values.description,
       createdBy: values.createdBy?.value,
@@ -161,6 +165,9 @@ export const VulnerabilityCreationForm: FunctionComponent<VulnerabilityFormProps
       confidence: parseInt(String(values.confidence), 10),
       externalReferences: values.externalReferences.map(({ value }) => value),
     };
+    if (values.file) {
+      finalValues.file = values.file;
+    }
     commit({
       variables: {
         input: finalValues,
@@ -278,6 +285,15 @@ export const VulnerabilityCreationForm: FunctionComponent<VulnerabilityFormProps
               style={fieldSpacingContainerStyle}
               setFieldValue={setFieldValue}
               values={values.externalReferences}
+          />
+          <Field
+            component={SimpleFileUpload}
+            name="file"
+            label={t('Associated file')}
+            FormControlProps={{ style: { marginTop: 20, width: '100%' } }}
+            InputLabelProps={{ fullWidth: true, variant: 'standard' }}
+            InputProps={{ fullWidth: true, variant: 'standard' }}
+            fullWidth={true}
           />
           <div className={classes.buttons}>
             <Button variant="contained"
