+++ opencti-platform/opencti-front/src/private/components/entities/events/EventCreation.tsx
@@ -1,30 +1,37 @@
-import React, { useState } from 'react';
+import React, { FunctionComponent, useState } from 'react';
 import { Field, Form, Formik } from 'formik';
 import Drawer from '@mui/material/Drawer';
 import Typography from '@mui/material/Typography';
 import Button from '@mui/material/Button';
 import IconButton from '@mui/material/IconButton';
 import Fab from '@mui/material/Fab';
-import MenuItem from '@mui/material/MenuItem';
 import { Add, Close } from '@mui/icons-material';
 import * as Yup from 'yup';
-import { graphql } from 'react-relay';
-import * as R from 'ramda';
+import { graphql, useMutation } from 'react-relay';
 import makeStyles from '@mui/styles/makeStyles';
+import { SimpleFileUpload } from 'formik-mui';
+import { RecordSourceSelectorProxy } from 'relay-runtime';
+import { FormikConfig } from 'formik/dist/types';
 import { useFormatter } from '../../../../components/i18n';
-import { commitMutation, handleErrorInForm } from '../../../../relay/environment';
+import { handleErrorInForm } from '../../../../relay/environment';
 import TextField from '../../../../components/TextField';
-import SelectField from '../../../../components/SelectField';
 import CreatedByField from '../../common/form/CreatedByField';
-import ObjectLabelField from '../../common/form/ObjectLabelField';
 import ObjectMarkingField from '../../common/form/ObjectMarkingField';
 import MarkDownField from '../../../../components/MarkDownField';
 import { ExternalReferencesField } from '../../common/form/ExternalReferencesField';
+import { parse } from '../../../../utils/Time';
+import DateTimePickerField from '../../../../components/DateTimePickerField';
+import OpenVocabField from '../../common/form/OpenVocabField';
+import { fieldSpacingContainerStyle } from '../../../../utils/field';
+import ObjectLabelField from '../../common/form/ObjectLabelField';
 import { useSchemaCreationValidation } from '../../../../utils/hooks/useEntitySettings';
 import { insertNode } from '../../../../utils/store';
-import { fieldSpacingContainerStyle } from '../../../../utils/field';
+import { Theme } from '../../../../components/Theme';
+import { Option } from '../../common/form/ReferenceField';
+import { EventCreationMutation, EventCreationMutation$variables } from './__generated__/EventCreationMutation.graphql';
+import { EventsLinesPaginationQuery$variables } from './__generated__/EventsLinesPaginationQuery.graphql';
 
-const useStyles = makeStyles((theme) => ({
+const useStyles = makeStyles<Theme>((theme) => ({
   drawerPaper: {
     minHeight: '100vh',
     width: '50%',
@@ -67,52 +74,106 @@ const useStyles = makeStyles((theme) => ({
   },
 }));
 
-const organizationMutation = graphql`
-  mutation OrganizationCreationMutation($input: OrganizationAddInput!) {
-    organizationAdd(input: $input) {
+const eventMutation = graphql`
+  mutation EventCreationMutation($input: EventAddInput!) {
+    eventAdd(input: $input) {
       id
       name
       description
       entity_type
       parent_types
-      ...OrganizationLine_node
+      ...EventLine_node
     }
   }
 `;
 
-export const OrganizationCreationForm = ({ updater, onReset, onCompleted,
-  defaultCreatedBy, defaultMarkingDefinitions }) => {
+interface EventAddInput {
+  name: string
+  description: string
+  event_types: string[],
+  start_time: Date | null,
+  stop_time: Date | null,
+  createdBy: Option | undefined
+  objectMarking: Option[]
+  objectLabel: Option[]
+  externalReferences: { value: string }[]
+  file: File | undefined
+}
+
+interface EventFormProps {
+  updater: (store: RecordSourceSelectorProxy, key: string) => void
+  onReset?: () => void;
+  onCompleted?: () => void;
+  defaultCreatedBy?: { value: string, label: string }
+  defaultMarkingDefinitions?: { value: string, label: string }[]
+}
+
+export const EventCreationForm: FunctionComponent<EventFormProps> = ({
+  updater,
+  onReset,
+  onCompleted,
+  defaultCreatedBy,
+  defaultMarkingDefinitions,
+}) => {
   const classes = useStyles();
   const { t } = useFormatter();
   const basicShape = {
     name: Yup.string().min(2).required(t('This field is required')),
     description: Yup.string().nullable(),
-    x_opencti_organization_type: Yup.string().nullable(),
-    x_opencti_reliability: Yup.string().nullable(),
+    event_types: Yup.array().nullable(),
+    start_time: Yup.date()
+      .typeError(t('The value must be a datetime (yyyy-MM-dd hh:mm (a|p)m)'))
+      .nullable(),
+    stop_time: Yup.date()
+      .typeError(t('The value must be a datetime (yyyy-MM-dd hh:mm (a|p)m)'))
+      .min(Yup.ref('start_time'), "The end date can't be before start date")
+      .nullable(),
   };
-  const organizationValidator = useSchemaCreationValidation('Organization', basicShape);
-  const onSubmit = (values, { setSubmitting, setErrors, resetForm }) => {
-    const finalValues = R.pipe(
-      R.assoc('createdBy', values.createdBy?.value),
-      R.assoc('objectMarking', R.pluck('value', values.objectMarking)),
-      R.assoc('objectLabel', R.pluck('value', values.objectLabel)),
-      R.assoc('externalReferences', R.pluck('value', values.externalReferences)),
-    )(values);
-    commitMutation({
-      mutation: organizationMutation,
+  const eventValidator = useSchemaCreationValidation('Event', basicShape);
+
+  const initialValues: EventAddInput = {
+    name: '',
+    description: '',
+    event_types: [],
+    start_time: null,
+    stop_time: null,
+    createdBy: defaultCreatedBy ?? '' as unknown as Option,
+    objectMarking: defaultMarkingDefinitions ?? [],
+    objectLabel: [],
+    externalReferences: [],
+    file: undefined,
+  };
+
+  const [commit] = useMutation<EventCreationMutation>(eventMutation);
+
+  const onSubmit: FormikConfig<EventAddInput>['onSubmit'] = (values, { setSubmitting, setErrors, resetForm }) => {
+    const finalValues: EventCreationMutation$variables['input'] = {
+      name: values.name,
+      description: values.description,
+      event_types: values.event_types,
+      start_time: values.start_time ? parse(values.start_time).format() : null,
+      stop_time: values.stop_time ? parse(values.stop_time).format() : null,
+      createdBy: values.createdBy?.value,
+      objectMarking: values.objectMarking.map((v) => v.value),
+      objectLabel: values.objectLabel.map((v) => v.value),
+      externalReferences: values.externalReferences.map(({ value }) => value),
+    };
+    if (values.file) {
+      finalValues.file = values.file;
+    }
+    commit({
       variables: {
         input: finalValues,
       },
       updater: (store) => {
         if (updater) {
-          updater(store, 'organizationAdd');
+          updater(store, 'eventAdd');
         }
       },
       onError: (error) => {
         handleErrorInForm(error, setErrors);
         setSubmitting(false);
       },
-      setSubmitting,
       onCompleted: () => {
         setSubmitting(false);
         resetForm();
@@ -123,18 +184,8 @@ export const OrganizationCreationForm = ({ updater, onReset, onCompleted,
     });
   };
 
-  return <Formik
-      initialValues={{
-        name: '',
-        description: '',
-        x_opencti_reliability: undefined,
-        x_opencti_organization_type: 'other',
-        createdBy: defaultCreatedBy ?? '',
-        objectMarking: defaultMarkingDefinitions ?? [],
-        objectLabel: [],
-        externalReferences: [],
-      }}
-      validationSchema={organizationValidator}
+  return <Formik initialValues={initialValues}
+      validationSchema={eventValidator}
       onSubmit={onSubmit}
       onReset={onReset}>
     {({
@@ -151,7 +202,15 @@ export const OrganizationCreationForm = ({ updater, onReset, onCompleted,
               name="name"
               label={t('Name')}
               fullWidth={true}
-              detectDuplicate={['Organization']}
+              detectDuplicate={['Event']}
+          />
+          <OpenVocabField
+              label={t('Event types')}
+              type="event-type-ov"
+              name="event_types"
+              containerStyle={fieldSpacingContainerStyle}
+              multiple={true}
+              onChange={(name, value) => setFieldValue(name, value)}
           />
           <Field
               component={MarkDownField}
@@ -159,39 +218,29 @@ export const OrganizationCreationForm = ({ updater, onReset, onCompleted,
               label={t('Description')}
               fullWidth={true}
               multiline={true}
-              rows="4"
+              rows={4}
               style={{ marginTop: 20 }}
           />
-          { /* TODO Improve customization (vocab with letter range) 2662 */}
           <Field
-              component={SelectField}
-              variant="standard"
-              name="x_opencti_organization_type"
-              label={t('Organization type')}
-              fullWidth={true}
-              containerstyle={fieldSpacingContainerStyle}
-          >
-            <MenuItem value="constituent">{t('Constituent')}</MenuItem>
-            <MenuItem value="csirt">{t('CSIRT')}</MenuItem>
-            <MenuItem value="partner">{t('Partner')}</MenuItem>
-            <MenuItem value="vendor">{t('Vendor')}</MenuItem>
-            <MenuItem value="other">{t('Other')}</MenuItem>
-          </Field>
+              component={DateTimePickerField}
+              name="start_time"
+              TextFieldProps={{
+                label: t('Start date'),
+                variant: 'standard',
+                fullWidth: true,
+                style: { marginTop: 20 },
+              }}
+          />
           <Field
-              component={SelectField}
-              variant="standard"
-              name="x_opencti_reliability"
-              label={t('Reliability')}
-              fullWidth={true}
-              containerstyle={fieldSpacingContainerStyle}
-          >
-            <MenuItem value="A">{t('reliability_A')}</MenuItem>
-            <MenuItem value="B">{t('reliability_B')}</MenuItem>
-            <MenuItem value="C">{t('reliability_C')}</MenuItem>
-            <MenuItem value="D">{t('reliability_D')}</MenuItem>
-            <MenuItem value="E">{t('reliability_E')}</MenuItem>
-            <MenuItem value="F">{t('reliability_F')}</MenuItem>
-          </Field>
+              component={DateTimePickerField}
+              name="stop_time"
+              TextFieldProps={{
+                label: t('End date'),
+                variant: 'standard',
+                fullWidth: true,
+                style: { marginTop: 20 },
+              }}
+          />
           <CreatedByField
               name="createdBy"
               style={fieldSpacingContainerStyle}
@@ -212,6 +261,15 @@ export const OrganizationCreationForm = ({ updater, onReset, onCompleted,
               style={fieldSpacingContainerStyle}
               setFieldValue={setFieldValue}
               values={values.externalReferences}
+          />
+          <Field
+            component={SimpleFileUpload}
+            name="file"
+            label={t('Associated file')}
+            FormControlProps={{ style: { marginTop: 20, width: '100%' } }}
+            InputLabelProps={{ fullWidth: true, variant: 'standard' }}
+            InputProps={{ fullWidth: true, variant: 'standard' }}
+            fullWidth={true}
           />
           <div className={classes.buttons}>
             <Button
@@ -237,7 +295,7 @@ export const OrganizationCreationForm = ({ updater, onReset, onCompleted,
   </Formik>;
 };
 
-const OrganizationCreation = ({ paginationOptions }) => {
+const EventCreation = ({ paginationOptions }: { paginationOptions: EventsLinesPaginationQuery$variables }) => {
   const classes = useStyles();
   const { t } = useFormatter();
   const [open, setOpen] = useState(false);
@@ -246,11 +304,11 @@ const OrganizationCreation = ({ paginationOptions }) => {
   const handleClose = () => setOpen(false);
   const onReset = () => handleClose();
 
-  const updater = (store) => insertNode(
+  const updater = (store: RecordSourceSelectorProxy) => insertNode(
     store,
-    'Pagination_organizations',
+    'Pagination_events',
     paginationOptions,
-    'organizationAdd',
+    'eventAdd',
   );
 
   return (
@@ -261,24 +319,26 @@ const OrganizationCreation = ({ paginationOptions }) => {
         className={classes.createButton}>
         <Add />
       </Fab>
-      <Drawer open={open}
+      <Drawer
+        open={open}
         anchor="right"
         elevation={1}
         sx={{ zIndex: 1202 }}
         classes={{ paper: classes.drawerPaper }}
         onClose={handleClose}>
         <div className={classes.header}>
-          <IconButton aria-label="Close"
+          <IconButton
+            aria-label="Close"
             className={classes.closeButton}
             onClick={handleClose}
             size="large"
             color="primary">
             <Close fontSize="small" color="primary" />
           </IconButton>
-          <Typography variant="h6">{t('Create an organization')}</Typography>
+          <Typography variant="h6">{t('Create an event')}</Typography>
         </div>
         <div className={classes.container}>
-          <OrganizationCreationForm
+          <EventCreationForm
               updater={updater}
               onCompleted={() => handleClose()}
               onReset={onReset}
@@ -289,4 +349,4 @@ const OrganizationCreation = ({ paginationOptions }) => {
   );
 };
 
-export default OrganizationCreation;
+export default EventCreation;
