+++ opencti-platform/opencti-front/src/private/components/entities/sectors/SectorCreation.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React, { FunctionComponent, useState } from 'react';
 import { Field, Form, Formik } from 'formik';
 import Drawer from '@mui/material/Drawer';
 import Typography from '@mui/material/Typography';
@@ -6,23 +6,32 @@ import Button from '@mui/material/Button';
 import IconButton from '@mui/material/IconButton';
 import Fab from '@mui/material/Fab';
 import { Add, Close } from '@mui/icons-material';
-import * as R from 'ramda';
 import * as Yup from 'yup';
-import { graphql } from 'react-relay';
+import { graphql, useMutation } from 'react-relay';
 import makeStyles from '@mui/styles/makeStyles';
+import { SimpleFileUpload } from 'formik-mui';
+import { RecordSourceSelectorProxy } from 'relay-runtime';
+import { FormikConfig } from 'formik/dist/types';
 import { useFormatter } from '../../../../components/i18n';
-import { commitMutation, handleErrorInForm } from '../../../../relay/environment';
+import { handleErrorInForm } from '../../../../relay/environment';
 import TextField from '../../../../components/TextField';
 import CreatedByField from '../../common/form/CreatedByField';
-import ObjectLabelField from '../../common/form/ObjectLabelField';
 import ObjectMarkingField from '../../common/form/ObjectMarkingField';
 import MarkDownField from '../../../../components/MarkDownField';
 import { ExternalReferencesField } from '../../common/form/ExternalReferencesField';
+import ObjectLabelField from '../../common/form/ObjectLabelField';
+import { fieldSpacingContainerStyle } from '../../../../utils/field';
 import { useSchemaCreationValidation } from '../../../../utils/hooks/useEntitySettings';
 import { insertNode } from '../../../../utils/store';
-import { fieldSpacingContainerStyle } from '../../../../utils/field';
+import { Theme } from '../../../../components/Theme';
+import { Option } from '../../common/form/ReferenceField';
+import {
+  SectorCreationMutation,
+  SectorCreationMutation$variables,
+} from './__generated__/SectorCreationMutation.graphql';
+import { SectorsLinesPaginationQuery$variables } from './__generated__/SectorsLinesPaginationQuery.graphql';
 
-const useStyles = makeStyles((theme) => ({
+const useStyles = makeStyles<Theme>((theme) => ({
   drawerPaper: {
     minHeight: '100vh',
     width: '50%',
@@ -65,50 +74,98 @@ const useStyles = makeStyles((theme) => ({
   },
 }));
 
-const systemMutation = graphql`
-  mutation SystemCreationMutation($input: SystemAddInput!) {
-    systemAdd(input: $input) {
+const sectorMutation = graphql`
+  mutation SectorCreationMutation($input: SectorAddInput!) {
+    sectorAdd(input: $input) {
       id
       name
       description
       entity_type
       parent_types
-      ...SystemLine_node
+      isSubSector
+      subSectors {
+        edges {
+          node {
+            id
+            name
+            description
+          }
+        }
+      }
     }
   }
 `;
 
-export const SystemCreationForm = ({ updater, onReset, onCompleted,
-  defaultCreatedBy, defaultMarkingDefinitions }) => {
+interface SectorAddInput {
+  name: string
+  description: string
+  createdBy: Option | undefined
+  objectMarking: Option[]
+  objectLabel: Option[]
+  externalReferences: { value: string }[]
+  file: File | undefined
+}
+
+interface SectorFormProps {
+  updater: (store: RecordSourceSelectorProxy, key: string) => void
+  onReset?: () => void;
+  onCompleted?: () => void;
+  defaultCreatedBy?: { value: string, label: string }
+  defaultMarkingDefinitions?: { value: string, label: string }[]
+}
+
+export const SectorCreationForm: FunctionComponent<SectorFormProps> = ({
+  updater,
+  onReset,
+  onCompleted,
+  defaultCreatedBy,
+  defaultMarkingDefinitions,
+}) => {
   const classes = useStyles();
   const { t } = useFormatter();
   const basicShape = {
     name: Yup.string().min(2).required(t('This field is required')),
     description: Yup.string().nullable(),
   };
-  const systemValidator = useSchemaCreationValidation('System', basicShape);
-  const onSubmit = (values, { setSubmitting, setErrors, resetForm }) => {
-    const finalValues = R.pipe(
-      R.assoc('createdBy', values.createdBy?.value),
-      R.assoc('objectMarking', R.pluck('value', values.objectMarking)),
-      R.assoc('objectLabel', R.pluck('value', values.objectLabel)),
-      R.assoc('externalReferences', R.pluck('value', values.externalReferences)),
-    )(values);
-    commitMutation({
-      mutation: systemMutation,
+  const sectorValidator = useSchemaCreationValidation('Sector', basicShape);
+
+  const initialValues: SectorAddInput = {
+    name: '',
+    description: '',
+    createdBy: defaultCreatedBy ?? '' as unknown as Option,
+    objectMarking: defaultMarkingDefinitions ?? [],
+    objectLabel: [],
+    externalReferences: [],
+    file: undefined,
+  };
+
+  const [commit] = useMutation<SectorCreationMutation>(sectorMutation);
+
+  const onSubmit: FormikConfig<SectorAddInput>['onSubmit'] = (values, { setSubmitting, setErrors, resetForm }) => {
+    const finalValues: SectorCreationMutation$variables['input'] = {
+      name: values.name,
+      description: values.description,
+      createdBy: values.createdBy?.value,
+      objectMarking: values.objectMarking.map((v) => v.value),
+      objectLabel: values.objectLabel.map((v) => v.value),
+      externalReferences: values.externalReferences.map(({ value }) => value),
+    };
+    if (values.file) {
+      finalValues.file = values.file;
+    }
+    commit({
       variables: {
         input: finalValues,
       },
       updater: (store) => {
         if (updater) {
-          updater(store, 'systemAdd');
+          updater(store, 'sectorAdd');
         }
       },
       onError: (error) => {
         handleErrorInForm(error, setErrors);
         setSubmitting(false);
       },
-      setSubmitting,
       onCompleted: () => {
         setSubmitting(false);
         resetForm();
@@ -119,19 +176,10 @@ export const SystemCreationForm = ({ updater, onReset, onCompleted,
     });
   };
 
-  return <Formik
-      initialValues={{
-        name: '',
-        description: '',
-        createdBy: defaultCreatedBy ?? '',
-        objectMarking: defaultMarkingDefinitions ?? [],
-        objectLabel: [],
-        externalReferences: [],
-      }}
-      validationSchema={systemValidator}
+  return <Formik initialValues={initialValues}
+      validationSchema={sectorValidator}
       onSubmit={onSubmit}
-      onReset={onReset}
-  >
+      onReset={onReset}>
     {({
       submitForm,
       handleReset,
@@ -146,7 +194,7 @@ export const SystemCreationForm = ({ updater, onReset, onCompleted,
               name="name"
               label={t('Name')}
               fullWidth={true}
-              detectDuplicate={['User']}
+              detectDuplicate={['Sector']}
           />
           <Field
               component={MarkDownField}
@@ -178,6 +226,15 @@ export const SystemCreationForm = ({ updater, onReset, onCompleted,
               setFieldValue={setFieldValue}
               values={values.externalReferences}
           />
+          <Field
+            component={SimpleFileUpload}
+            name="file"
+            label={t('Associated file')}
+            FormControlProps={{ style: { marginTop: 20, width: '100%' } }}
+            InputLabelProps={{ fullWidth: true, variant: 'standard' }}
+            InputProps={{ fullWidth: true, variant: 'standard' }}
+            fullWidth={true}
+          />
           <div className={classes.buttons}>
             <Button
                 variant="contained"
@@ -202,18 +259,18 @@ export const SystemCreationForm = ({ updater, onReset, onCompleted,
   </Formik>;
 };
 
-const SystemCreation = ({ paginationOptions }) => {
+const SectorCreation = ({ paginationOptions }: { paginationOptions: SectorsLinesPaginationQuery$variables }) => {
   const classes = useStyles();
   const { t } = useFormatter();
   const [open, setOpen] = useState(false);
 
   const handleOpen = () => setOpen(true);
   const handleClose = () => setOpen(false);
-  const updater = (store) => insertNode(
+  const updater = (store: RecordSourceSelectorProxy) => insertNode(
     store,
-    'Pagination_systems',
+    'Pagination_sectors',
     paginationOptions,
-    'systemAdd',
+    'sectorAdd',
   );
 
   return (
@@ -221,7 +278,8 @@ const SystemCreation = ({ paginationOptions }) => {
         <Fab onClick={handleOpen}
           color="secondary"
           aria-label="Add"
-          className={classes.createButton}>
+          className={classes.createButton}
+        >
           <Add />
         </Fab>
         <Drawer open={open}
@@ -240,14 +298,14 @@ const SystemCreation = ({ paginationOptions }) => {
             >
               <Close fontSize="small" color="primary" />
             </IconButton>
-            <Typography variant="h6">{t('Create a system')}</Typography>
+            <Typography variant="h6">{t('Create a sector')}</Typography>
           </div>
           <div className={classes.container}>
-            <SystemCreationForm updater={updater} onCompleted={handleClose} onReset={handleClose}/>
+            <SectorCreationForm updater={updater} onCompleted={handleClose} onReset={handleClose}/>
           </div>
         </Drawer>
       </div>
   );
 };
 
-export default SystemCreation;
+export default SectorCreation;
