+++ opencti-platform/opencti-front/src/private/components/observations/stix_cyber_observables/StixCyberObservableCreation.js
@@ -62,6 +62,7 @@ import { fieldSpacingContainerStyle } from '../../../../utils/field';
 import { insertNode } from '../../../../utils/store';
 import { useFormatter } from '../../../../components/i18n';
 import useVocabularyCategory from '../../../../utils/hooks/useVocabularyCategory';
+import { SimpleFileUpload } from "formik-mui";
 
 const useStyles = makeStyles((theme) => ({
   drawerPaper: {
@@ -340,6 +341,9 @@ const StixCyberObservableCreation = ({
         obsContent: values.obsContent?.value,
       },
     };
+    if (values.file) {
+      finalValues.file = values.file;
+    }
     commitMutation({
       mutation: stixCyberObservableMutation,
       variables: finalValues,
@@ -421,6 +425,7 @@ const StixCyberObservableCreation = ({
               objectLabel: [],
               externalReferences: [],
               createIndicator: false,
+              file: undefined,
             };
             const attributes = pipe(
               map((n) => n.node),
@@ -615,6 +620,15 @@ const StixCyberObservableCreation = ({
                       style={fieldSpacingContainerStyle}
                       setFieldValue={setFieldValue}
                       values={values.externalReferences}
+                    />
+                    <Field
+                      component={SimpleFileUpload}
+                      name="file"
+                      label={t('Associated file')}
+                      FormControlProps={{ style: { marginTop: 20, width: '100%' } }}
+                      InputLabelProps={{ fullWidth: true, variant: 'standard' }}
+                      InputProps={{ fullWidth: true, variant: 'standard' }}
+                      fullWidth={true}
                     />
                     <Field
                       component={SwitchField}
