+++ opencti-platform/opencti-front/src/private/components/techniques/narratives/NarrativeCreation.tsx
@@ -1,5 +1,4 @@
-import React, { useState } from 'react';
-import * as R from 'ramda';
+import React, { FunctionComponent, useState } from 'react';
 import { Field, Form, Formik } from 'formik';
 import Drawer from '@mui/material/Drawer';
 import Typography from '@mui/material/Typography';
@@ -8,13 +7,16 @@ import IconButton from '@mui/material/IconButton';
 import Fab from '@mui/material/Fab';
 import { Add, Close } from '@mui/icons-material';
 import * as Yup from 'yup';
-import { graphql } from 'react-relay';
+import { graphql, useMutation } from 'react-relay';
 import Dialog from '@mui/material/Dialog';
 import DialogTitle from '@mui/material/DialogTitle';
 import DialogContent from '@mui/material/DialogContent';
 import makeStyles from '@mui/styles/makeStyles';
+import { SimpleFileUpload } from 'formik-mui';
+import { RecordSourceSelectorProxy } from 'relay-runtime';
+import { FormikConfig } from 'formik/dist/types';
 import { useFormatter } from '../../../../components/i18n';
-import { commitMutation, handleErrorInForm } from '../../../../relay/environment';
+import { handleErrorInForm } from '../../../../relay/environment';
 import TextField from '../../../../components/TextField';
 import CreatedByField from '../../common/form/CreatedByField';
 import ObjectLabelField from '../../common/form/ObjectLabelField';
@@ -23,9 +25,16 @@ import MarkDownField from '../../../../components/MarkDownField';
 import { ExternalReferencesField } from '../../common/form/ExternalReferencesField';
 import { useSchemaCreationValidation } from '../../../../utils/hooks/useEntitySettings';
 import { insertNode } from '../../../../utils/store';
+import { Theme } from '../../../../components/Theme';
+import { Option } from '../../common/form/ReferenceField';
+import {
+  NarrativeCreationMutation,
+  NarrativeCreationMutation$variables,
+} from './__generated__/NarrativeCreationMutation.graphql';
+import { NarrativesLinesPaginationQuery$variables } from './__generated__/NarrativesLinesPaginationQuery.graphql';
 import { fieldSpacingContainerStyle } from '../../../../utils/field';
 
-const useStyles = makeStyles((theme) => ({
+const useStyles = makeStyles<Theme>((theme) => ({
   drawerPaper: {
     minHeight: '100vh',
     width: '50%',
@@ -96,8 +105,39 @@ const narrativeMutation = graphql`
   }
 `;
 
-export const NarrativeCreationForm = ({ updater, onReset, inputValue, onCompleted,
-  defaultCreatedBy, defaultMarkingDefinitions }) => {
+interface NarrativeAddInput {
+  name: string
+  description: string
+  confidence: number
+  createdBy: Option | undefined
+  objectMarking: Option[]
+  objectLabel: Option[]
+  externalReferences: { value: string }[]
+  file: File | undefined
+}
+
+interface NarrativeFormProps {
+  updater: (store: RecordSourceSelectorProxy, key: string) => void
+  paginationOptions?: NarrativesLinesPaginationQuery$variables;
+  display?: boolean;
+  contextual?: boolean;
+  onReset?: () => void;
+  inputValue?: string;
+  onCompleted?: () => void;
+  defaultCreatedBy?: { value: string, label: string }
+  defaultMarkingDefinitions?: { value: string, label: string }[]
+  defaultConfidence?: number;
+}
+
+export const NarrativeCreationForm: FunctionComponent<NarrativeFormProps> = ({
+  updater,
+  onReset,
+  inputValue,
+  onCompleted,
+  defaultCreatedBy,
+  defaultMarkingDefinitions,
+  defaultConfidence,
+}) => {
   const classes = useStyles();
   const { t } = useFormatter();
   const basicShape = {
@@ -105,15 +145,34 @@ export const NarrativeCreationForm = ({ updater, onReset, inputValue, onComplete
     description: Yup.string().nullable(),
   };
   const narrativeValidator = useSchemaCreationValidation('Narrative', basicShape);
-  const onSubmit = (values, { setSubmitting, setErrors, resetForm }) => {
-    const finalValues = R.pipe(
-      R.assoc('createdBy', values.createdBy?.value),
-      R.assoc('objectMarking', R.pluck('value', values.objectMarking)),
-      R.assoc('objectLabel', R.pluck('value', values.objectLabel)),
-      R.assoc('externalReferences', R.pluck('value', values.externalReferences)),
-    )(values);
-    commitMutation({
-      mutation: narrativeMutation,
+
+  const initialValues: NarrativeAddInput = {
+    name: inputValue ?? '',
+    description: '',
+    confidence: defaultConfidence ?? 75,
+    createdBy: defaultCreatedBy ?? '' as unknown as Option,
+    objectMarking: defaultMarkingDefinitions ?? [],
+    objectLabel: [],
+    externalReferences: [],
+    file: undefined,
+  };
+
+  const [commit] = useMutation<NarrativeCreationMutation>(narrativeMutation);
+
+  const onSubmit: FormikConfig<NarrativeAddInput>['onSubmit'] = (values, { setSubmitting, setErrors, resetForm }) => {
+    const finalValues: NarrativeCreationMutation$variables['input'] = {
+      name: values.name,
+      description: values.description,
+      confidence: parseInt(String(values.confidence), 10),
+      createdBy: values.createdBy?.value,
+      objectMarking: values.objectMarking.map((v) => v.value),
+      objectLabel: values.objectLabel.map((v) => v.value),
+      externalReferences: values.externalReferences.map(({ value }) => value),
+    };
+    if (values.file) {
+      finalValues.file = values.file;
+    }
+    commit({
       variables: {
         input: finalValues,
       },
@@ -126,7 +185,6 @@ export const NarrativeCreationForm = ({ updater, onReset, inputValue, onComplete
         handleErrorInForm(error, setErrors);
         setSubmitting(false);
       },
-      setSubmitting,
       onCompleted: () => {
         setSubmitting(false);
         resetForm();
@@ -138,14 +196,7 @@ export const NarrativeCreationForm = ({ updater, onReset, inputValue, onComplete
   };
 
   return <Formik
-      initialValues={{
-        name: inputValue ?? '',
-        description: '',
-        createdBy: defaultCreatedBy ?? '',
-        objectMarking: defaultMarkingDefinitions ?? [],
-        objectLabel: [],
-        externalReferences: [],
-      }}
+      initialValues={initialValues}
       validationSchema={narrativeValidator}
       onSubmit={onSubmit}
       onReset={onReset}>
@@ -195,6 +246,15 @@ export const NarrativeCreationForm = ({ updater, onReset, inputValue, onComplete
                     setFieldValue={setFieldValue}
                     values={values.externalReferences}
                 />
+                <Field
+                  component={SimpleFileUpload}
+                  name="file"
+                  label={t('Associated file')}
+                  FormControlProps={{ style: { marginTop: 20, width: '100%' } }}
+                  InputLabelProps={{ fullWidth: true, variant: 'standard' }}
+                  InputProps={{ fullWidth: true, variant: 'standard' }}
+                  fullWidth={true}
+                />
                 <div className={classes.buttons}>
                     <Button
                         variant="contained"
@@ -217,14 +277,19 @@ export const NarrativeCreationForm = ({ updater, onReset, inputValue, onComplete
     </Formik>;
 };
 
-const NarrativeCreation = ({ paginationOptions, contextual, inputValue, display }) => {
+const NarrativeCreation: FunctionComponent<NarrativeFormProps> = ({
+  paginationOptions,
+  contextual,
+  inputValue,
+  display,
+}) => {
   const classes = useStyles();
   const { t } = useFormatter();
   const [open, setOpen] = useState(false);
 
   const handleOpen = () => setOpen(true);
   const handleClose = () => setOpen(false);
-  const updater = (store) => insertNode(
+  const updater = (store: RecordSourceSelectorProxy) => insertNode(
     store,
     'Pagination_narratives',
     paginationOptions,
