+++ opencti-platform/opencti-graphql/src/modules/administrativeArea/administrativeArea.graphql
@@ -190,6 +190,7 @@ input AdministrativeAreaAddInput {
     modified: DateTime
     clientMutationId: String
     update: Boolean
+    file: Upload
 }
 
 type Mutation {
