+++ opencti-platform/opencti-graphql/src/modules/channel/channel.graphql
@@ -190,6 +190,7 @@ input ChannelAddInput {
     modified: DateTime
     clientMutationId: String
     update: Boolean
+    file: Upload
 }
 
 type Mutation {
