+++ opencti-platform/opencti-graphql/src/modules/dataComponent/dataComponent.graphql
@@ -195,6 +195,7 @@ input DataComponentAddInput {
     description: String
     aliases: [String]
     dataSource: String
+    file: Upload
 }
 
 type Mutation {
