+++ opencti-platform/opencti-graphql/src/modules/event/event.graphql
@@ -193,6 +193,7 @@ input EventAddInput {
     modified: DateTime
     clientMutationId: String
     update: Boolean
+    file: Upload
 }
 
 type Mutation {
