+++ opencti-platform/opencti-graphql/src/modules/grouping/grouping.graphql
@@ -252,6 +252,7 @@ input GroupingAddInput {
     modified: DateTime
     clientMutationId: String
     update: Boolean
+    file: Upload
 }
 
 type Mutation {
