+++ opencti-platform/opencti-graphql/src/modules/narrative/narrative.graphql
@@ -190,6 +190,7 @@ input NarrativeAddInput {
     modified: DateTime
     clientMutationId: String
     update: Boolean
+    file: Upload
 }
 
 type Mutation {
